requirements.txt
fastapi
pinatapy-vourhey
firebase_admin
uvicorn[standard]
uuid